Extract some cursor util from debug module
