+++ src/extension-test/unit/cursor-doc/utilities-test.ts
@@ -1,5 +1,6 @@
 import * as expect from 'expect';
 import * as utilities from '../../../cursor-doc/utilities';
+import { docFromTextNotation } from '../common/text-notation';
 //import { docFromTextNotation, textAndSelection } from '../common/text-notation';
 
 describe('Utilities that need cursor doc things', () => {
@@ -32,6 +33,18 @@ describe('Utilities that need cursor doc things', () => {
       const text = '(a b) (c {:d \n; ( [ ( {\n[1 2 3 "four"';
       const append = ']})';
       expect(utilities.getMissingBrackets(text)).toEqual({ append: append, prepend: '' });
+    });
+  });
+  describe('isRightSexpStructural', () => {
+    it('map', () => {
+      const doc = docFromTextNotation('|{:a 1}');
+      const cursor = doc.getTokenCursor(0);
+      expect(utilities.isRightSexpStructural(cursor)).toBe(true);
+    });
+    it('Tagged map', () => {
+      const doc = docFromTextNotation('|#foo {:a 1}');
+      const cursor = doc.getTokenCursor(0);
+      expect(utilities.isRightSexpStructural(cursor)).toBe(true);
     });
   });
 });
